Render JFnet applicable to larger images
All FC layers converted to ConvLayer equivalents.
Reshape Layers have to be adopted accordingly to
work with the new tensor dimensionalities.
For dealing with other than 512**2 images, untie_biases
has to be set to False.
Because softmax does not work on 4D tensors, we
temporarily concatenate multiple spatial predictions
along the batch dim and then reshape back again.

Changes to be committed:
	modified:   models.py