Test correct output of JFnet
passes for the current JFnet that can be applied
to larger images and produces the same output
as the net with dense layers (original implementation),
two commits before. To be more specific, it produces
the same output for 512 images and untied_biases.
Tied_biases seem to change the posteriors a bit.

 Changes to be committed:
	modified:   tests/test_models.py